Reduce available states
